requirements.txt
# Requires Python >=3.10,<3.12 for maximum compatibility
fastapi==0.110.2
uvicorn[standard]==0.27.1
SQLAlchemy>=2.0
asyncpg==0.29.0
alembic>=1.13.0
databases[postgresql]>=0.8.0
psycopg2-binary>=2.9.0
pydantic==2.6.4
pydantic-settings==2.1.0
python-dotenv==1.0.1
python-docx==1.1.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
requests>=2.31.0